refactor: refactor upload service file structure
