+++ src/upload/index.ts
@@ -6,10 +6,10 @@ import config from 'config'
 import { Application, UploadService, ServiceAddresses } from '../definitions'
 import { loggingFactory } from '../logger'
 import { errorHandler, waitForReadyApp } from '../utils'
-import UploadJob from './upload.model'
-import uploadHandler from './upload.handler'
-import getFileSizeHandler from './getFileSize.handler'
-import getSizeLimitHandler from './getSizeLimit.handler'
+import UploadJob from './model/upload.model'
+import uploadHandler from './handler/upload.handler'
+import getFileSizeHandler from './handler/getFileSize.handler'
+import getSizeLimitHandler from './handler/getSizeLimit.handler'
 
 export const UPLOAD_FOLDER = 'uploads'
 const logger = loggingFactory('upload-service')
