+++ test/integration/upload.spec.ts
@@ -14,7 +14,7 @@ import { rooms } from '../../src/communication'
 import { MessageCodesEnum, ServiceAddresses, UploadJobStatus } from '../../src/definitions'
 import { ProviderManager } from '../../src/providers'
 import { UPLOAD_FOLDER } from '../../src/upload'
-import UploadJob from '../../src/upload/upload.model'
+import UploadJob from '../../src/upload/model/upload.model'
 import { duplicateObject } from '../../src/utils'
 import jobsGC from '../../src/gc'
 import { createLibp2pRoom, sleep, spawnLibp2p } from '../utils'
